Shots now destroy asteroids and corrected some type hints

+++ sprites/asteroid.py
@@ -2,11 +2,11 @@ import pygame
 from sprites.circleshape import CircleShape
 
 class Asteroid(CircleShape):
-    def __init__(self, x: int, y: int, radius: int):
+    def __init__(self, x: float, y: float, radius: float):
         super().__init__(x, y, radius)
 
     def draw(self, screen: pygame.display):
         pygame.draw.circle(screen, "white", self.position, self.radius, 2)
 
-    def update(self, dt: int):
+    def update(self, dt: float):
         self.position = self.position + (self.velocity * dt)
